fixed fuzzy intersection calculation
This will avoid counting similar tokens if they've already been
(partially) counted. Also, for all non-crisp intersection types,
this removes more mass from the non-intersection portion of
partially matched strings. The "matched" weight remains the same
as before, but twice as much "matched" weight is removed from
the non-intersection portion. This is to account for partial
matching only giving each part of a partial match 1/2 of the
"matched" weight.